% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustools.R
\name{filter1_params}
\alias{filter1_params}
\title{Default parameters for the first filtering step}
\usage{
filter1_params(dataset)
}
\arguments{
\item{dataset}{Name of the toy dataset (either "quake", "sandberg",
"linnarsson" or "bernstein")}
}
\value{
min.cells, max.cells, min.reads required to be able to filter genes
using filter1.
}
\description{
Returns default min.cells, max.cells, min.reads parameters used in
gene_filter1()
}
\examples{
filter1_params("quake")
}

