% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustools.R
\name{norm_laplacian}
\alias{norm_laplacian}
\title{Laplacian computation}
\usage{
norm_laplacian(x, tau)
}
\arguments{
\item{x}{Adjacency/distance matrix}

\item{tau}{Regularization term}
}
\value{
Laplacian of the adjacency/distance matrix
}
\description{
Laplacian computation
}
\examples{
L <- norm_laplacian(dists, 0)
}

