% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustools.R
\name{transformation}
\alias{transformation}
\title{Dimensionality reduction of a distance matrix}
\usage{
transformation(dists, method)
}
\arguments{
\item{dists}{A distance matrix}

\item{method}{Dimensionality reduction method: "pca", "spectral", "spectral_reg"
or "mds"}
}
\value{
A transformed distance matrix
}
\description{
Transform a distance matrix to a new basis
}
\examples{
transformation(d, "spectral")
}

